requirements.txt
flask
pillow
piexif